chore: break post deployment steps into separate files
