+++ scripts/post_deployment_step_10.js
@@ -61,26 +61,6 @@ async function main() {
     const sBrickArtifact = await get('sOlympus');
     const sBrickAddress = sBrickArtifact.address;
     await treasury.toggle('9', sBrickAddress, zeroAddress);
-
-    const stakingArtifact = await get('OlympusStaking');
-
-    const sBrick = (
-      await ethers.getContractFactory('sOlympus')
-    ).attach(sBrickArtifact.address);
-
-    const currentStakingContract = await sBrick.stakingContract();
-    if (currentStakingContract === config.contractAddresses.zero) {
-      await sBrick.initialize(stakingArtifact.address);
-    }
-
-    // TODO: What should index be?
-    // Copied from SQUID for now
-    // https://etherscan.io/tx/0x927640e4d8fb17f859472ec9c54b8c0f6ebe8ec1f1747c61447111dc49185019
-    const currentIndex = await sBrick.INDEX();
-    if (currentIndex.eq(0)) {
-      const index = 1000000000;
-      await sBrick.setIndex(index);
-    }
 }
 
 main()
